Review every code comment against the code it describes
Four independent reviews read every comment of the project's code — the
three kernels and their shared queue and buffers, the Cortex-M layer
and the RP ports, the STM32 port, the emulation, the tests, the tools
and the CI — against the code, and each finding was checked again
before it was applied. Only comments, docstrings, Robot documentation,
one CI step name and two CI error messages, translated to English,
changed: the 98 images of the 21 builds of the Pico, Pico 2 and F4
examples are identical byte for byte, and the host tests, the three
models and the F4 Renode suites pass.

Among what was wrong: EDF*'s tie-break, which the header described the
other way round from the code; event-driven tasks called a sporadic
server where the formula is a total bandwidth server (Spuri and
Buttazzo, 1996); the source of an interrupt said to be masked while it
is served, as on the MSP430 and not on Cortex-M; the bootrom said to
start the SRAM image, which the debugger does; the RP2040 vector table
placed at 0x20000000 and the RP2350 timer's interrupt registers four
words away instead of two; header examples that would not compile;
references to _OSTime, removed files and the MSP430; "// Claude" as the
comment of a field; the host test's clock, the F4 wrap platform's
frequency and the cause of a reboot, in measure_cost.sh, all stated
wrongly. Missing reasons were added where the code depends on them:
the LL/SC reservation lost to any interrupt, the layout a queue
descriptor shares with FIFOQUEUE, the table of the 3-slot writer.

Code defects the reviews met are left for commits of their own:
trace_gpio.sh no longer sees falling edges since the examples write
BSRR in 32 bits, the DEBUG_MODE priority check never fires, the RP2040
UART does not wait for IO_BANK0 out of reset, FIFO nodes are only
2-byte aligned, and a few in dead or STM32-only code.

+++ Escapement/CORTEX-Mx/STM32/Escapement_TimerEvent.h
@@ -35,7 +35,7 @@
 ** their occurrence time.
 ** Parameters:
 **  (1) (UINT8) nbNode: maximum number of pending events;
-**  (2) (UINT8) prescaler: divides the counter clock frequency, i.e. the timer frequency
+**  (2) (UINT16) prescaler: divides the counter clock frequency, i.e. the timer frequency
 **      will be set to f(input)/(prescaler+1).
 **  (3) (UINT8) priority: priority level of the timer;
 **  (4) (UINT8) subpriority: subpriority level of the timer;
